Change all sh files and corrects labels

+++ .rm_dir.sh
@@ -5,7 +5,22 @@ LCYAN='\033[1;36m'
 
 current_dir=$(pwd)
 
-echo -e "${LCYAN}|Delete folder done!|${BREAK}"
+for i in $(ls ~); do
+    if [ "${i}" == "Desktop" ]; then 
+        DESK=${i}
+        break
+    elif [ "${i}" == "Рабочий стол" ]; then
+        DESK=${i}
+        break
+    else 
+        DESK="0"
+    fi
+done
 
-rm /home/$USER/Desktop/WordEng.desktop
-rm -rf $current_dir
+if [ "${DESK}" != "0" ]; then 
+    echo -e "${LCYAN}|Delete folder done!|${BREAK}"
+    rm /home/$USER/${DESK}/WordEng.desktop
+    rm -rf $current_dir
+else
+    echo -e "${LCYAN}I NOT FOUND DESKTOP${BREAK}"  
+fi
